drm-core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "drm-core"
version = "0.1.0"
description = "Core DRM Library for GPU Management"
requires-python = ">=3.8"
dependencies = [
    "aiohttp>=3.8.0",
    "sqlalchemy>=1.4.0",
    "python-dotenv>=0.19.0",
    "tabulate>=0.8.0",
    "ray[default]>=2.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]
dev = [
    "pre-commit>=3.0.0"
]

[tool.hatch.build.targets.wheel]
packages = ["src/drm_core"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
